calculate percentile of observed data
The percentile is now added to the field data frame. I calculated the percentile position of the observed data relative to the shuffled distribution in each bar. This will be used as a p value in future steps after corrections for multiple comparisons.

I used scipy's percentileofscore function on the normalized (to the time spent in the bar) shuffled data and the observed normalised data.