Increase CI swap space to overcome ram limits
The default GitHub Actions servers allocate very little ram
to us. This causes the integration tests to fail.

To get around this, we'll allocate some disk space as backup ram
(aka swap space). This can be increased if necessary but will
need to make sure that we still have enough disk space
to hold the data and any temporary files.