change firing rate threshold to 0
I changed how the mean firing rate is calculated so we stop combining
data from multiple recordings in spatial_firing (the mean firing rate
was calculated from all of them). To avoid missing silent cells I'm changing
this threshold to 0, but I'm open to discussing this / just setting it
for myself. It seems possible to pick up cells with such low rates in my
data when the recordings are long enough so there's no reason to discard them.