Update firing_maps.py
In lines 49,50, and 97,98, the px and py are calculated as: x*bin_size + (bin_size/2), where x or y is bin number. The bin_size_cm or bin_size_pixels in either of those places has the same value of 2.5, and given the formula the bins would be counted as follows: 0-1.25, 1.25-3.75, 3.75-6.25, ...., 96.25-98.75. But they should be 0-2.5, 2.5-5, 5-7.5, ...97.5-100. Basically the bins are offset by 1.25 cms. But x*bin_size + bin_size would remove this offset. 
In short: current formula x*m + m/2. Proposed formula: x*m + m.