fix clusterid mismatch error
The curation metrics were calculated for the combined recordings and
had all cluster ids in them. Some recordings have clusters that only
fired in one session and not the other. When the curation metrics were
loaded, this function assumed that cluster ids were consecutive and
loaded incorrect metrics for some recordings (for example when a cluster
should have been skipped).